同步F0的工程，优化exti

+++ ebox/hardware/cortex-m0/mcu.cpp
@@ -28,7 +28,7 @@
 #define systick_interrupt()     SysTick->CTRL |=0x0002
 
 extern "C" {
-    __IO uint64_t millis_seconds;//提供一个mills()等效的全局变量。降低cpu调用开销
+    __IO uint32_t millis_seconds;//提供一个mills()等效的全局变量。降低cpu调用开销
     __IO uint16_t micro_para;
 
     static void update_system_clock(CpuClock_t *clock);
@@ -84,9 +84,9 @@ extern "C" {
       *@param    mcu
       *@retval   none
       */
-    uint64_t mcu_micros(void)
+    uint32_t mcu_micros(void)
     {
-        uint64_t micro;
+        uint32_t micro;
         //        uint32_t temp = __get_PRIMASK();//保存之前中断设置
         //        no_interrupts();
         //        if (SysTick->CTRL & (1 << 16))//发生了溢出
@@ -105,7 +105,7 @@ extern "C" {
       *@param    mcu
       *@retval   none
       */
-    uint64_t mcu_millis( void )
+    uint32_t mcu_millis( void )
     {
         return millis_seconds;
     }
